requirements.txt
grpcio==1.38.1
grpcio-tools==1.38.1
deepspeech==0.9.3
PyAudio==0.2.11
